# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/HelloHandler": {
            "get": {
                "description": "这是第一个API",
                "tags": [
                    "文章"
                ],
                "summary": "第一个API",
                "responses": {
                    "200": {},
                    "400": {}
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "通过邮箱获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户的邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功后返回值",
                        "schema": {
                            "$ref": "#/definitions/dao.User"
                        }
                    },
                    "400": {
                        "description": "fail\" \"失败后返回值",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "用户"
                ],
                "summary": "新建用户",
                "parameters": [
                    {
                        "description": "用户",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/user.CreateUserParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功后返回值",
                        "schema": {
                            "$ref": "#/definitions/dao.User"
                        }
                    },
                    "400": {
                        "description": "fail\" \"失败后返回值",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dao.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "user.CreateUserParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}